Re-use the stream implementation for Crypto.
